requirements.txt
# ── Core ──────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Data Collection ───────────────────────────────────────
pytrends>=4.9.2          # Google Trends (relative values)
httpx>=0.27.0            # async HTTP client (competitor crawling)
beautifulsoup4>=4.12.0   # HTML parsing
lxml>=5.2.0              # BS4 parser backend

# ── Google Search Console API ─────────────────────────────
google-api-python-client>=2.131.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0

# ── Prediction (dummy phase) ──────────────────────────────
numpy>=1.26.0
scikit-learn>=1.4.0      # linear regression dummy predictor

# ── Claude API ────────────────────────────────────────────
anthropic>=0.28.0

# ── Storage ───────────────────────────────────────────────
# sqlite3 is stdlib — no install needed

# ── CLI & Utilities ───────────────────────────────────────
rich>=13.7.0             # pretty terminal output
tenacity>=8.3.0          # retry logic with backoff
